refactor: モード選択画面をリファクタリング #175
