+++ front/src/app/presentation/SelectMode/SelectModePresentation.module.scss
@@ -40,45 +40,3 @@
     flex-direction: row;
   }
 }
-
-.modeSection {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 16px;
-
-  @media screen and (min-width: 1024px) {
-    gap: 24px;
-  }
-}
-
-.comment {
-  font-size: 16px;
-  width: 320px;
-  text-align: center;
-
-  @media screen and (min-width: 1024px) {
-    font-size: 20px;
-    width: 440px;
-  }
-}
-
-.button {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 8px;
-
-  @media screen and (min-width: 1024px) {
-    gap: 16px;
-  }
-}
-
-.icon {
-  height: 4rem;
-  width: auto;
-}
