+++ front/src/features/select-mode/components/ButtonArea/ButtonArea.module.scss
@@ -1,0 +1,17 @@
+.button {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
+
+  @media screen and (min-width: 1024px) {
+    gap: 16px;
+  }
+}
+
+.icon {
+  height: 4rem;
+  width: auto;
+}
