+++ front/src/features/select-mode/components/CommentArea/CommentArea.module.scss
@@ -1,0 +1,10 @@
+.comment {
+  font-size: 16px;
+  width: 320px;
+  text-align: center;
+
+  @media screen and (min-width: 1024px) {
+    font-size: 20px;
+    width: 440px;
+  }
+}
