+++ front/src/features/select-mode/components/IdeaMemoReviewSection/IdeaMemoReviewSection.module.scss
@@ -1,0 +1,30 @@
+.modeSection {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 16px;
+
+  @media screen and (min-width: 1024px) {
+    gap: 24px;
+  }
+}
+
+.button {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
+
+  @media screen and (min-width: 1024px) {
+    gap: 16px;
+  }
+}
+
+.icon {
+  height: 4rem;
+  width: auto;
+}
