+++ front/src/features/select-mode/components/StartIdeaSessionSection/StartIdeaSessionSection.module.scss
@@ -1,0 +1,12 @@
+.modeSection {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 16px;
+
+  @media screen and (min-width: 1024px) {
+    gap: 24px;
+  }
+}
