+++ front/src/features/select-mode/components/index.ts
@@ -1,0 +1,4 @@
+import IdeaMemoReviewSection from "@/features/select-mode/components/IdeaMemoReviewSection/IdeaMemoReviewSection";
+import StartIdeaSessionSection from "@/features/select-mode/components/StartIdeaSessionSection/StartIdeaSessionSection";
+
+export { IdeaMemoReviewSection, StartIdeaSessionSection };
